+++ core/Factory.py
@@ -1,8 +1,8 @@
 from battle.Battle import BattleCog, ToonAttack
 from core.Cog import Cog
 from core.Entity import BattleEntity, Entity
+from core.Toon import Toon
 from gags.Gag import Gag, get_gag_min_max_exp
-from Toon import Toon
 
 from .GagGlobals import TRACK
