+++ tests/inventory/test_inventory_creation.py
@@ -1,7 +1,7 @@
 from core.GagGlobals import DEFAULT_GAG_COUNT, DEFAULT_GAG_LIMIT
 from core.ToonGlobals import DEFAULT_BEAN_COUNT, DEFAULT_BEAN_LIMIT
 from gags.Gag import DEFAULT_TRACK_EXPS_CURRENT, Gags, get_default_gags
-from Toon import Inventory
+from core.Toon import Inventory
 
 
 class TestInventoryDefaultCreation:
